instantsplat/requirements.txt
camtools==0.1.4
einops==0.7.0
lpips==0.1.4
tensorboard==2.15.0
tqdm==4.66.1
transformers==4.35.2
omegaconf==2.3.0
# open3d==0.18.0
opencv-python==4.8.1.78
Pillow==10.0.1
plyfile==1.0.1
pytorch-lightning==2.1.0
PyYAML==6.0.1
ipywidgets
jupyterlab
sql
